pyproject.toml
[project]
name = "ComfyUI-AbstractImaGen"
description = "ComfyUI node for [a/Abstract Image Generation](https://github.com/wakattac/ComfyUI-AbstractImaGen/). This node is designed to create unique abstract base images on the fly within your ComfyUI workflows, which can then be used as input for VAE encoding, image-to-image generation, or other creative processes."
version = "2.0.0"
license = { file = "LICENSE" }
dependencies = ["torch", "pillow", "numpy", "scipy"]

[project.urls]
Repository = "https://github.com/wakattac/ComfyUI-AbstractImaGen"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "wakattac"
DisplayName = "ComfyUI-AbstractImaGen"
Icon = ""
